modified notifications

+++ src/main/java/com/example/littleloomlocator/model/NotificationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
 
-	List<Notification> findBySenderId(int senderId);
-	List<Notification> findByReceiverId(int receiverId);
+	List<Notification> findBySenderId(long senderId);
+	List<Notification> findByReceiverId(long receiverId);
 }
